Create Structure
To help ascertain the relationship between the spectral bands and the predictions, we present our decision tree models’ structure in text form. The model was constructed with training samples from different dates, collected at the riparian zone of the Parana river. The presented structure is from the pruned tree model, with a total of 394 leaves and 787 nodes.